moving the launch darkly msw handlers to the mocks package
