+++ lib/packages/shared-types/opensearch/main/transforms/toggle-withdraw-rai.ts
@@ -1,5 +1,4 @@
-// import { toggleWithdrawRaiEnabledSchema } from "../../..";
-import { events } from "shared-types";
+import { events } from "../../../index";
 
 // transform used to take in id: string
 export const transform = () => {
