+++ test/package.json
@@ -7,7 +7,9 @@
   "main": "index.js",
   "scripts": {
     "e2e": "playwright test --project=ci",
-    "e2e:ui": "playwright test --ui --project=ci"
+    "e2e:ui": "playwright test --ui --project=ci",
+    "e2e:local": "playwright test --project=local",
+    "e2e:local:ui": "playwright test --ui --project=local"
   },
   "author": "",
   "license": "ISC",
